chore: Update constants and routers
This commit updates the `constants.py` file to include a new constant `DEVSERVER_DOMAIN` and adds it to the `main.py` file. It also removes unused batch scripts and updates the README.md file. Additionally, it includes changes to the routers, such as adding the `application_router` to `osu_client/__init__.py` and updating the prefix for `api_authentication_router` to "/api/v1/authentication".